# app/static/swagger.yaml
swagger: "2.0"
info:
  title: "Mechanic API"
  description: "This API is used to create, update, and delete information for a mechanic shop"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: |
      Token Required!
      Please enter your JWT token **with `Bearer` prefix** like:
      ```
      Bearer "JWT TOKEN"
      ```
paths:
...
  /tickets:
    post:
      tags:
        - Service Tickets
      summary: "Create a new service ticket"
      description:
        Creates a new service ticket in the system. Requires admin authentication and supports assigning multiple service types.
        If `is_major_damage` is true, the ticket will also be sent to a sister site.
        Validates all `service_type_ids` provided. If any ID is invalid (doesn't exist), returns a 400 with a list of invalid IDs.
      consumes:
        - application/json
      produces:
        - application/json
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          description: JSON payload to create the service ticket (with optional service_type_ids array)
          required: true
          schema:
            type: object
            required:
              - service_date
              - customer_id
              - car_id
              - VIN
              - car_issue
              - is_major_damage
            properties:
              service_date:
                type: string
                format: date
                example: "2025-05-25"
              customer_id:
                type: integer
              car_id:
                type: integer
              VIN:
                type: string
                description: "The vehicle identification number"
              car_issue:
                type: string
                description: "The issue reported with the car"
              is_major_damage:
                type: boolean
              service_type_ids:
                type: array
                items:
                  type: integer
                description: List of service type ID's to associate with the ticket
      responses:
        201:
          description: "Successfully created the ticket"
          schema:
            type: object
            properties:
              message:
                type: string
              ticket:
                $ref: "#/definitions/CreateServiceTicketPayload"
          examples:
            application/json:
              message: "Ticket created"
              ticket:
                id: 10
                service_date: "2025-05-25"
                customer_id: 1
                car_id: 2
                VIN: "1HGCM82633A004352"
                car_issue: "Brakes not working"
                is_major_damage: false
                services:
                  - id: 1
                    name: "Oil Change"
        400:
          description: Bad Request - Validation Error
          schema:
            type: object
            properties:
              message:
                type: string
              invalid_ids:
                type: array
                items:
                  type: integer
          example:
            message: "Invalid service_type_ids"
            invalid_ids: [888, 1234]
        401:
          description: Unauthorized - Token missing or invalid
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: Forbidden - Admin access required
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    get:
      tags:
        - Service Tickets
      summary: "Gets all tickets currently in the system"
      description: "Returns a list of all tickets with details on what is being worked on and what car it is assigned to"
      parameters:
        - name: page
          in: query
          type: integer
          required: false
          description: "Page number (default: 1)"
        - name: per_page
          in: query
          type: integer
          required: false
          description: "Number of results per page (default: 10)"
      responses:
        200:
          description: "Retrieved all tickets successfully"
          schema:
            $ref: "#/definitions/AllTickets"
          examples:
            application/json:
              tickets:
                - id: 1
                  service_date: "2025-06-01"
                  customer_id: 1
                  car_id: 1
                  VIN: 123
                  car_issue: "Oil Leak"
                  is_major_damage: false
                  services:
                    - id: 1
                      name: "Oil Change"
                - id: 2
                  service_date: "2025-05-25"
                  customer_id: 2
                  car_id: 2
                  VIN: 234
                  car_issue: "Tire Fix"
                  is_major_damage: false
                  services:
                    - id: 2
                      name: "Tire Replacement"
              page: 1
              per_page: 10
              total_tickets: 2
  /tickets/{id}:
    get:
      tags:
        - Service Tickets
      summary: "Gets tickets by ID"
      description: "Returns full ticket details including assigned services, customer and car IDs, and damage info"
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          description: "The ID of the service ticket to retrieve"
      responses:
        200:
          description: Ticket retrieved successfully
          schema:
            $ref: "#/definitions/ServiceTicketPayload"
          examples:
            application/json:
              id: 10
              service_date: "2025-05-25"
              customer_id: 1
              car_id: 2
              VIN: "1HGCM82633A004352"
              car_issue: "Brake failure"
              is_major_damage: false
              services:
                - id: 1
                  name: "Oil Change"
        401:
          description: Unauthorized - Token missing or invalid
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Ticket not found
          schema:
            $ref: "#/definitions/ErrorResponse"

    put:
      tags:
        - Service Tickets
      summary: "Update an existing service ticket"
      description: |
        Updates an existing ticket’s data. Requires admin authentication.
        If `service_type_ids` is provided, validates all IDs and associates them with the ticket.
      security:
        - bearerAuth: []
      consumes:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          description: "The ID of the service ticket to update"
        - in: body
          name: body
          description: Updated ticket payload
          required: true
          schema:
            type: object
            required:
              - service_date
              - customer_id
              - car_id
              - VIN
              - car_issue
              - is_major_damage
            properties:
              service_date:
                type: string
                format: date
                example: "2025-06-01"
              customer_id:
                type: integer
              car_id:
                type: integer
              VIN:
                type: string
              car_issue:
                type: string
              is_major_damage:
                type: boolean
              service_type_ids:
                type: array
                items:
                  type: integer
                description: Optional list of service type IDs to update associations
      responses:
        200:
          description: Successfully updated ticket
          schema:
            type: object
            properties:
              message:
                type: string
              ticket:
                $ref: "#/definitions/ServiceTicketPayload"
          examples:
            application/json:
              message: "Ticket updated"
              ticket:
                id: 10
                service_date: "2025-06-01"
                customer_id: 1
                car_id: 2
                VIN: "1HGCM82633A004352"
                car_issue: "Brake pads replaced"
                is_major_damage: false
                services:
                  - id: 2
                    name: "Brake Replacement"
        400:
          description: Bad Request - Validation Error or Invalid service_type_ids
          schema:
            type: object
            properties:
              message:
                type: string
              invalid_ids:
                type: array
                items:
                  type: integer
        401:
          description: Unauthorized - Token missing or invalid
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: Forbidden - Admin access required
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Ticket not found
          schema:
            $ref: "#/definitions/ErrorResponse"

    delete:
      tags:
        - Service Tickets
      summary: "Endpoint to delete a ticket"
      description: "Deletes a service ticket. Requires token authentication, employee must be logged in to delete the ticket"
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          description: "ID of the ticket to delete"
      responses:
        200:
          description: "Ticket successfully deleted"
          schema:
            $ref: "#/definitions/DeleteTicketResponse"
          examples:
            application/json:
              message: "Successfully deleted ticket <ticket_id>"
        401:
          description: Unauthorized - Missing or invalid token
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: Forbidden - Admin token required
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Ticket not found
          schema:
            $ref: "#/definitions/ErrorResponse"
  /tickets/{ticket_id}/assign-mechanic/{employee_id}:
    put:
      tags:
        - Service Tickets
      summary: "Assign a mechanic to a ticket"
      description: "Assign an employee (mechanic) to a specific service ticket. Requires admin token."
      security:
        - bearerAuth: []
      parameters:
        - name: ticket_id
          in: path
          required: true
          type: integer
          description: "ID of the ticket"
        - name: employee_id
          in: path
          required: true
          type: integer
          description: "ID of the mechanic to assign"
      responses:
        200:
          description: "Mechanic assigned to ticket successfully"
          schema:
            type: object
            properties:
              message:
                type: string
          examples:
            application/json:
              message: "Employee 3 assign to ticket 7."
        401:
          description: Unauthorized - Missing or invalid token
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: Forbidden - Admin access required
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Ticket or Employee not found
          schema:
            $ref: "#/definitions/ErrorResponse"
  /tickets/{ticket_id}/remove-mechanic/{employee_id}:
    put:
      tags:
        - Service Tickets
      summary: "Remove a mechanic from a ticket"
      descriptions: "Removes an assigned mechanic from a service ticket. Requires Admin Token"
      security:
        - bearerAuth: []
      parameters:
        - name: ticket_id
          in: path
          required: true
          type: integer
          description: "ID of the service ticket"
        - name: employee_id
          in: path
          required: true
          type: integer
          description: "ID of the mechanic to remove"
      responses:
        200:
          description: "Mechanic removed from ticket successfully"
          schema:
            type: object
            properties:
              message:
                type: string
          examples:
            application/json:
              message: "Employee 3 removed from ticket 7."
        401:
          description: Unauthorized - Missing or invalid token
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: Forbidden - Admin access required
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Ticket or Employee not found
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  ErrorResponse:
    type: "object"
    properties:
      message:
        type: "string"
...
  CreateServiceTicketPayload:
    type: object
    properties:
      id:
        type: integer
      VIN:
        type: string
      car_issue:
        type: string
      service_date:
        type: string
        format: date
      customer_id:
        type: integer
      car_id:
        type: integer
      is_major_damage:
        type: boolean
      services:
        type: array
        items:
          $ref: "#/definitions/ServiceType"
  AllTickets:
    type: "object"
    properties:
      tickets:
        type: "array"
        items:
          $ref: "#/definitions/CreateTicketResponse"
      page:
        type: "integer"
      per_page:
        type: "integer"
      total_tickets:
        type: "integer"
  CreateTicketResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      service_date:
        type: "string"
        format: "date"
      customer_id:
        type: "integer"
      car_id:
        type: "integer"
      VIN:
        type: "string"
      car_issue:
        type: "string"
      is_major_damage:
        type: "boolean"
      services:
        type: array
        items:
          $ref: "#/definitions/ServiceType"
  DeleteTicketResponse:
    type: "object"
    properties:
      message:
        type: string
        description: Confirmation message of ticket deletion
...
  ServiceType:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
...